pyproject.toml
[project]
name = "tethyr"
version = "0.1.0"
description = "AR Processing Server with Gemini Live integration for smart glasses"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "websockets>=15.0.1",
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "loguru>=0.7.3",
    "scipy>=1.16.2",
    "ultralytics>=8.3.221",
    "opencv-python>=4.12.0.88",
    "opencv-contrib-python>=4.12.0.88",
    "clip",
    "google-genai[aiohttp]>=1.46.0",
    "livekit-agents[google]~=1.2",
]


[dependency-groups]
dev = ["ruff>=0.8.4", "pre-commit>=4.0.1"]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 150
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.uv.sources]
clip = { git = "https://github.com/ultralytics/CLIP.git" }
